Update use-kms-etcd-encryption.md
Encrypting Secret Data at Rest ---> Using a KMS provider for data encryption

The title of the linked kubernetes.io page is  "Using a KMS provider for data encryption". Not "Encrypting Secret Data at Rest".

Or, if the title "Encrypting Secret Data at Rest" is supposed to be correct, then the link URL should be changed.